+++ internal/api/handlers/common/get_version_test.go
@@ -6,6 +6,7 @@ import (
 
 	"allaboutapps.dev/aw/go-starter/internal/api"
 	"allaboutapps.dev/aw/go-starter/internal/test"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -14,5 +15,19 @@ func TestGetVersion(t *testing.T) {
 		res := test.PerformRequest(t, s, "GET", "/-/version?mgmt-secret="+s.Config.Management.Secret, nil, nil)
 		require.Equal(t, http.StatusOK, res.Result().StatusCode)
 		require.Equal(t, "build.local/misses/ldflags @ < 40 chars git commit hash via ldflags > (1970-01-01T00:00:00+00:00)", res.Body.String()) // build args are not injected during test time.
+
+		// caching: ensure this call is always uncached for browsers
+		assert.Equal(t, "no-cache, private, max-age=0", res.Header().Get("Cache-Control"))
+		assert.Equal(t, "Thu, 01 Jan 1970 00:00:00 UTC", res.Header().Get("Expires"))
+		assert.Equal(t, "0", res.Header().Get("X-Accel-Expires"))
+		assert.Equal(t, "no-cache", res.Header().Get("Pragma"))
+
+		// caching: unset
+		assert.Equal(t, "", res.Header().Get("ETag"))
+		assert.Equal(t, "", res.Header().Get("If-Modified-Since"))
+		assert.Equal(t, "", res.Header().Get("If-Match"))
+		assert.Equal(t, "", res.Header().Get("If-None-Match"))
+		assert.Equal(t, "", res.Header().Get("If-Range"))
+		assert.Equal(t, "", res.Header().Get("If-Unmodified-Since"))
 	})
 }