+++ internal/api/router/router.go
@@ -147,7 +147,7 @@ func Init(s *api.Server) {
 		// Unsecured base group available at /**
 		Root: s.Echo.Group(""),
 
-		// Management endpoints, secured by key auth (query param), available at /-/**
+		// Management endpoints, uncacheable, secured by key auth (query param), available at /-/**
 		Management: s.Echo.Group("/-", echoMiddleware.KeyAuthWithConfig(echoMiddleware.KeyAuthConfig{
 			KeyLookup: "query:mgmt-secret",
 			Validator: func(key string, c echo.Context) (bool, error) {
@@ -160,7 +160,7 @@ func Init(s *api.Server) {
 				}
 				return false
 			},
-		})),
+		}), middleware.NoCache()),
 
 		// OAuth2, unsecured or secured by bearer auth, available at /api/v1/auth/**
 		APIV1Auth: s.Echo.Group("/api/v1/auth", middleware.AuthWithConfig(middleware.AuthConfig{